feat(model-decorator): make model options with table name required
Because of some issues when uglifying code for browser usage we could not provide a consistent behavior for the table name if none was provided using the options. The correct information to define the table name got lost after uglifying the code.
Therefor the model options are now required to be provided with a table name.

BREAKING CHANGE
opts with tableName is now required when decorating a Model with @Model(opts: ModelData). This makes sure we have consistent behavior also when running uglified code.